feat(form): add cfField label.js for config handling of mandatory asterisk
