+++ packages/form/addon/components/cf-field.hbs
@@ -13,10 +13,7 @@
     {{in-viewport onEnter=this.refreshDynamicOptions}}
   >
     {{#if this.labelVisible}}
-      <CfField::Label
-        @field={{@field}}
-        @useMandatoryAsterisk={{this.useMandatoryAsterisk}}
-      />
+      <CfField::Label @field={{@field}} />
     {{/if}}
 
     <div class="uk-flex">
